pyproject.toml
[tool.poetry]
name = "elk-experiments"
version = "0.1.0"
description = ""
authors = ["Oliver Daniels-Koch <danielskoch.oliver@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
ipykernel = "^6.29.4"
huggingface-hub = {version = "^0.23.0", extras = ["cli"]}
torch = "~2.0"
transformers = "^4.41.1"
datasets = "^2.19.1"
cupbearer = {path = "cupbearer", develop = true}
einops = "^0.8.0"
accelerate = "^0.30.1"
transformer-lens = "^2.0.0"
blobfile = "^2.1.1"
ipywidgets = "^8.1.3"
ipdb = "^0.13.13"
omegaconf = "^2.3.0"
hydra-core = "^1.3.2"
hydra-submitit-launcher = "^1.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
